feat: add card internationalization (#82)
I added card internationalization (i18n) to the cards by adding a separate card translation entity. This allows future expansion on supported languages. The language of each card translation should be a ISO 639-1 code. You can specify a language on the request or leave it empty to get the default. Using a '*' will return everything not flattened. Furthermore, I added the ability to filter on tags. Moreover, the test coverage is now at 100% for this application.

There is a performance caveat with this implementation, because some filtering and flattening is done on the business logic level and not on the database level.